feat: 헤더 옵션 늘어남에 따라 반응형 컴포넌트 및 애니메이션 추가
