+++ src/style/header.css
@@ -42,24 +42,24 @@ menu {
 
   @media screen and (max-width: 799px) {
     display: none;
+    width: 100%;
+    position: absolute;
+    flex-direction: column;
+    background-color: #1d212a;
+    margin-top: 23rem;
+    opacity: 0.8;
+    transform: translateX(100%);
+    transition: all 0.5s cubic-bezier(0.25, 0.8, 0.25, 1);
+
+    li {
+      font-size: 2rem;
+      text-align: right;
+      padding-right: 4rem;
+    }
   }
 }
 .__opened {
-  width: 100%;
-  position: absolute;
   display: flex;
-  flex-direction: column;
-  background-color: #1d212a;
-  margin-top: 23rem;
-  opacity: 0.8;
-  transform: translateX(100%);
-  transition: all 0.5s cubic-bezier(0.25, 0.8, 0.25, 1);
-
-  li {
-    font-size: 2rem;
-    text-align: right;
-    padding-right: 4rem;
-  }
 }
 
 #nav__menu__button {
